TRUNK-5816 Exclude powermock from api
since we cannot run PowerMock tests on JUnit 5. This way we can already
prevent new contributions from adding PowerMock based tests. Since we
are in a transition period where we most likely will not be able to get
rid of the JUnit 4 dependency. Because we need to provide the JUnit 4
based BaseContext... classes to module developers. We need to ensure
that no new JUnit 4 tests are added to the openmrs-core